+++ src/Commands/Views/CreateFormViewCommand.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace CrestApps\CodeGenerator\Commands;
+namespace CrestApps\CodeGenerator\Commands\Views;
 
 use CrestApps\CodeGenerator\Commands\Bases\ViewsCommandBase;
 use CrestApps\CodeGenerator\Models\Resource;
 
-class CreateEditViewCommand extends ViewsCommandBase
+class CreateFormViewCommand extends ViewsCommandBase
 {
     /**
      * The name and signature of the console command.
      *
      * @var string
      */
-    protected $signature = 'create:edit-view
+    protected $signature = 'create:form-view
                             {model-name : The model name that this view will represent.}
                             {--resource-file= : The name of the resource-file to import from.}
                             {--views-directory= : The name of the directory to create the views under.}
@@ -27,7 +27,7 @@ class CreateEditViewCommand extends ViewsCommandBase
      *
      * @var string
      */
-    protected $description = 'Create an edit-views for the model.';
+    protected $description = 'Create a from-view for the model.';
 
     /**
      * Gets the name of the stub to process.
@@ -36,7 +36,7 @@ class CreateEditViewCommand extends ViewsCommandBase
      */
     protected function getStubName()
     {
-        return 'edit.blade';
+        return 'form.blade';
     }
 
     /**
@@ -52,41 +52,28 @@ class CreateEditViewCommand extends ViewsCommandBase
 
         if ($this->canCreateView($destenationFile, $input->force, $resources)) {
             $stub = $this->getStub();
+            $htmlCreator = $this->getHtmlGenerator($resources->fields, $input->modelName, $this->getTemplateName());
+            $headers = $this->getHeaderFieldAccessor($resources->fields, $input->modelName);
 
             $this->createLanguageFile($input->languageFileName, $input->resourceFile, $input->modelName)
-                ->createMissingViews($input)
                 ->replaceCommonTemplates($stub, $input, $resources->fields)
-                ->replaceFileUpload($stub, $resources->fields)
-                ->replacePrimaryKey($stub, $this->getPrimaryKeyName($resources->fields))
-                ->replaceModelHeader($stub, $this->getHeaderFieldAccessor($resources->fields, $input->modelName))
-                ->replaceFormId($stub, $this->getFormId($input->modelName))
-                ->replaceFormName($stub, $this->getFormName($input->modelName))
+                ->replaceFields($stub, $htmlCreator->getHtmlFields())
+                ->replaceModelHeader($stub, $headers)
                 ->createFile($destenationFile, $stub)
-                ->info('Edit view was crafted successfully.');
+                ->info('Form view was crafted successfully.');
         }
     }
 
     /**
-     * Gets te create form name
+     * Replaces the form field's html code in a giving stub.
      *
-     * @param string $modelName
+     * @param string $stub
+     * @param string $fields
      *
-     * @return string
+     * @return $this
      */
-    protected function getFormName($modelName)
+    protected function replaceFields(&$stub, $fields)
     {
-        return sprintf('edit_%s_form', snake_case($modelName));
-    }
-
-    /**
-     * Gets te create form id
-     *
-     * @param string $modelName
-     *
-     * @return string
-     */
-    protected function getFormId($modelName)
-    {
-        return sprintf('edit_%s_form', snake_case($modelName));
+        return $this->replaceTemplate('form_fields_html', $fields, $stub);
     }
 }